FCM test

+++ src/main/java/stg/onyou/model/network/FcmMessage.java
@@ -34,7 +34,7 @@ public class FcmMessage {
     @AllArgsConstructor
     @Getter
     public static class Data {
-        private Long clubId;
+        private Integer clubId;
     }
 
 }
